refactored request methods
